refs #1866,
tol_bvmat_bsr.h:168:37: error cast from 'const chat*' to 'int' loses precision

git-svn-id: https://www.tol-project.org/svn/tolp/trunk@6796 be4846ae-f051-483a-831f-b64e7d08527f

+++ tol/btol/vmatrix_type/tol_bvmat_bsr.h
@@ -165,7 +165,7 @@ struct missing_info
       "\n  index="+index+", "+
       "\n  row="+row+", "+
       "\n  col="+col+", "+
-      "\n  prior="+(int)prior.c_str()+", "+
+      "\n  prior="+prior.c_str()+", "+
       "\n  nu="+nu+", "+
       "\n  sigma2="+sigma2+", "+
       "\n  minBound="+minBound+", "+
